Some test fixes for latest `gufe` changes
